+++ apps/console/src/features/private-key-jwt/api/private-key-jwt.ts
@@ -1,0 +1,55 @@
+/**
+ * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { HttpMethods } from "@wso2is/core/models";
+import useRequest, {
+    RequestConfigInterface,
+    RequestErrorInterface,
+    RequestResultInterface
+} from "../../../features/core/hooks/use-request";
+
+/**
+ * Get JTI reuse state
+ *
+ * @returns The response of the JTI reuse state.
+ */
+export const useTokenReuseConfigData = <Data = any, Error = RequestErrorInterface>(
+): RequestResultInterface<Data, Error> => {
+
+    const requestConfig: RequestConfigInterface = {
+        headers: {
+            "Accept": "application/json",
+            "Content-Type": "application/json"
+        },
+        method: HttpMethods.GET,
+        params: {},
+        url: "https://run.mocky.io/v3/b1742d0d-5832-402e-afb7-80abf321b871"
+    };
+
+    const { data, error, isValidating, mutate } = useRequest<Data, Error>(requestConfig, {
+        shouldRetryOnError: false
+    });
+
+    return {
+        data,
+        error: error,
+        isLoading: !data && !error,
+        isValidating,
+        mutate
+    };
+};
